Move models from slackhealthbot.services.models to slackhealthbot.core.models.
